// src/app/store/auth/types.ts
export enum ACTION_TYPES {
    LOGIN_REQUEST = 'AUTH/LOGIN_REQUEST',
    LOGIN_REQUEST_SUCCESS = 'AUTH/LOGIN_REQUEST_SUCCESS',
    LOGIN_REQUEST_FAIL = 'AUTH/LOGIN_REQUEST_FAIL',
    FACEBOOK_LOGIN_REQUEST = 'AUTH/FACEBOOK_LOGIN_REQUEST',
    FACEBOOK_LOGIN_REQUEST_SUCCESS = 'AUTH/FACEBOOK_LOGIN_REQUEST_SUCCESS',
    FACEBOOK_LOGIN_REQUEST_FAIL = 'AUTH/FACEBOOK_LOGIN_REQUEST_FAIL',
    SET_USER = 'AUTH/SET_USER',
    SET_LOADING_USER = 'AUTH/SET_LOADING_USER',
}